bugfix: restrict_between_alignment_positions needs to restrict all nodes, except gap-only leaves because these are later removed from the tree
For gap-only internal nodes, it now sets a genomic_align that has the
correct "alignment" and "sequence" lengths, aligned and original
sequences, and cigar-line. As per the convention of
`GenomicAlign::restrict`, dnafrag_start is dnafrag_end+1, which is a bit
weird, but that's what allows the "sequence length" to be 0.

The Ortheus Runnable is now aware of that, and uses existing methods to
split the trees in subtrees that have non-empty ancestral sequences and
at least 2 leaves, whilst removing the gap-only columns that could have
appeared.